Share AnimePulse oauth handler + register in userscript entry (parity with other providers)
